refactor bindings into seperate files
